Rationalize the ACT running average computation in the custom dynesty sampler.
* No longer issue a warning when there are repeated likelihood values in 
the sampler output.  

* Add `adapt_tscale`: the running average ACT estimate will average over 
this number of livepoint replacements using an exponentially-decaying 
weight.  Default is 100.

* Get aggressive with `nact = 1` by default now that the ACT computation 
is stabilized.  Each MCMC live point replacement will run for one 
estimated ACT.  This means, on average, two accepted MCMC steps, so a 
~10% (= exp(-2)) chance of zero acceptances and a repeated live point.  
This does not bias the sampler (see http://arxiv.org/abs/1805.03924).

* The ACT estimate is an exponentially-weighted running mean of 
2/P_accept - 1 over each live point replacement; 2/P_accept - 1 is the 
ACT of an ideal rejection sampler with acceptance rate P_accept, so is a 
lower-bound for the ACT of the MCMC.